Fix celery restart

+++ install.sh
@@ -393,7 +393,9 @@ launch_celery(){
         echo '## Restart Celery ##'
         kill $( cat "$destfull"probemanager/celery.pid)
         pkill -f celery
-        rm "$destfull"probemanager/celery.pid
+        if [ -f "$destfull"probemanager/celery.pid ]; then
+            rm "$destfull"probemanager/celery.pid
+        fi
         sleep 8
         (cd "$destfull"probemanager/ && "$destfull"venv/bin/celery -A probemanager worker -D --pidfile celery.pid -B -l info -f /var/log/probemanager-celery.log --scheduler django_celery_beat.schedulers:DatabaseScheduler)
     fi
